Trying to submit working-but-unfinished code, to move towards more often git com mits.
Make it possible for BatchedLookup callers to tell if a given object was updated.

Only store most significant digit for number-of-invitees to an event. This signi
ficantly reduces the number of FBEvent/DBEvent/searchindex updates as a result.
Hack in need of cleanup!

Add an incremental-indexer to handle changes to facebook events, only when event is changed.

For search index updates, in addition to fb-event-info changes, also handle the case of recent events without changes, if they change from FUTURE to PAST, so we update the DB and search index. This is a horrible hack, and needs a different approach to fixing it.