Replace all our old manual regex-concatenation rules with the new rules.py hotness. Fix some bugs with rules.py discovered in the process of using them and comparing them against our historical dataset in classifiers/tools/eval_auto_classifier.py. Added a test for one that took a bit of time to discover (the dance_battles rule wasn't also including good_dance_battles rule).
I've hacked up regex_keywords a bit in the process of doing this and supporting both. But now that all rules and keywords are converted over, I'm hopeful that I can change how I approach the actual keyword detection in text, and clean up our regex generation and storage, and regex_keywords in that process.

I'm also hopeful that the new rules and keyword setup will bring a ton of clarity to the system in terms of operating at a higher level, and make it easier to understand our logic/matching, as well as extend it to create new rules easily when I rebuild my testing framework to focus on that again.