Add a proper admin site (appengine's admin is lacking important features like repeated-field editing, json editing, etc).
This uses flask-admin site (which needs my recent patch to support appengine datastore backend), as well as a patch to wtforms-appengine (to support ndb.JsonProperty)
https://github.com/mikelambert/flask-admin/commit/909a0f93f56de14e4a3488c2f491508f797bc009
https://github.com/mikelambert/wtforms-appengine/commit/131d3d82a117579557a8490263b70f9f7ba92e1e
For now these are not integrated upstream yet, so they exist only in my local repo. If they don't get released and pushed, I'll have to make this point directly to my github repositories instead.

This also imports flask (which looks like a promising direction to move the rest of the site to), so we'll probably move in that direction, given the level of support libraries built around flask.