Add a mapreduce to find access tokens that work for an event.
It adds a servlet, that given a list of event ids, iterates over all users, searching for which users have access to these events. It uses the new map-and-reduce mapreduces, so sticks many jobs onto the queues. We should try to call this with as many event_ids as possible.