Pull out the react-native-locale getCurrentLocale funtionality into the mobile/ client.
Turns out that react-native-locale has an index.ios.js and index.android.js but no index.js, so the webpack packager was unable to find the module for import (since it doesn't know about the per-platform conventions). This will probably come back to bite us with other modules in the future, and so we'll need to be sure to extract such dependencies from the common/server code.

Starting to get a feel for how this common development will go... :)